requirements.txt
streamlit
pandas
matplotlib.pyplot
base64
numpy
plotly_express
